chore(engine): day 5 — move smart-rule matcher to @node9/policy-engine
matchesPattern, evaluateSmartConditions, getNestedValue (rules/) and the
ReDoS-safe regex utilities (utils/regex.ts) are now in the engine.
Proxy re-exports keep existing import paths working. Engine version of
getCompiledRegex drops the NODE9_DEBUG console.error — engine stays
process-free; callers already handle the null return.
